requirements/prod.txt
gunicorn==19.0.0
boto==2.31.0
django-storages==1.1.8
